pointers

+++ kurl_util/cmd/config/main.go
@@ -107,51 +107,53 @@ func processSelinuxConfig(installer *kurlv1beta1.Installer, execCmds bool, gener
 	}
 
 	deleteScript := true
-	if generateScript && (installer.Spec.SelinuxConfig.Selinux != "" || installer.Spec.SelinuxConfig.Type != "") {
-		scriptLines := []string{
-			"BYPASS_SELINUX_PREFLIGHT=1",
-		}
-
-		switch installer.Spec.SelinuxConfig.Selinux {
-		case "enforcing":
-			scriptLines = append(scriptLines, "sed -i s/^SELINUX=.*$/SELINUX=enforcing/ /etc/selinux/config")
-		case "permissive":
-			scriptLines = append(scriptLines, "setenforce 0")
-			scriptLines = append(scriptLines, "sed -i s/^SELINUX=.*$/SELINUX=permissive/ /etc/selinux/config")
-		case "disabled":
-			scriptLines = append(scriptLines, "setenforce 0")
-			scriptLines = append(scriptLines, "sed -i s/^SELINUX=.*$/SELINUX=disabled/ /etc/selinux/config")
-		case "":
-		default:
-			return errors.Errorf("unknown selinux option: %s", installer.Spec.SelinuxConfig.Selinux)
-		}
-		if installer.Spec.SelinuxConfig.Type != "" {
-			line := fmt.Sprintf("sed -i s/^SELINUXTYPE=.*$/SELINUXTYPE=%s/ /etc/selinux/config", installer.Spec.SelinuxConfig.Type)
-			scriptLines = append(scriptLines, line)
-		}
-		if installer.Spec.SelinuxConfig.Selinux == "enforcing" {
-			// this always has to be at the end of the script or all other commands will fail
-			scriptLines = append(scriptLines, "setenforce 1")
-		}
-
-		script := fmt.Sprintf("configure_selinux() {\n\t%s\n}", strings.Join(scriptLines, "\n\t"))
-		if err := writeScript(scriptFilename, script); err != nil {
-			return errors.Wrap(err, "faied to save script")
-		}
-		deleteScript = false
-	}
-
-	if execCmds {
-		for _, args := range installer.Spec.SelinuxConfig.ChconCmds {
-			err := runCommand("chcon", args)
-			if err != nil {
-				return errors.Wrap(err, "failed to run chcon")
+	if installer.Spec.SelinuxConfig != nil {
+		if generateScript && (installer.Spec.SelinuxConfig.Selinux != "" || installer.Spec.SelinuxConfig.Type != "") {
+			scriptLines := []string{
+				"BYPASS_SELINUX_PREFLIGHT=1",
 			}
+
+			switch installer.Spec.SelinuxConfig.Selinux {
+			case "enforcing":
+				scriptLines = append(scriptLines, "sed -i s/^SELINUX=.*$/SELINUX=enforcing/ /etc/selinux/config")
+			case "permissive":
+				scriptLines = append(scriptLines, "setenforce 0")
+				scriptLines = append(scriptLines, "sed -i s/^SELINUX=.*$/SELINUX=permissive/ /etc/selinux/config")
+			case "disabled":
+				scriptLines = append(scriptLines, "setenforce 0")
+				scriptLines = append(scriptLines, "sed -i s/^SELINUX=.*$/SELINUX=disabled/ /etc/selinux/config")
+			case "":
+			default:
+				return errors.Errorf("unknown selinux option: %s", installer.Spec.SelinuxConfig.Selinux)
+			}
+			if installer.Spec.SelinuxConfig.Type != "" {
+				line := fmt.Sprintf("sed -i s/^SELINUXTYPE=.*$/SELINUXTYPE=%s/ /etc/selinux/config", installer.Spec.SelinuxConfig.Type)
+				scriptLines = append(scriptLines, line)
+			}
+			if installer.Spec.SelinuxConfig.Selinux == "enforcing" {
+				// this always has to be at the end of the script or all other commands will fail
+				scriptLines = append(scriptLines, "setenforce 1")
+			}
+
+			script := fmt.Sprintf("configure_selinux() {\n\t%s\n}", strings.Join(scriptLines, "\n\t"))
+			if err := writeScript(scriptFilename, script); err != nil {
+				return errors.Wrap(err, "faied to save script")
+			}
+			deleteScript = false
 		}
-		for _, args := range installer.Spec.SelinuxConfig.SemanageCmds {
-			err := runCommand("semanage", args)
-			if err != nil {
-				return errors.Wrap(err, "failed to run semanage")
+
+		if execCmds {
+			for _, args := range installer.Spec.SelinuxConfig.ChconCmds {
+				err := runCommand("chcon", args)
+				if err != nil {
+					return errors.Wrap(err, "failed to run chcon")
+				}
+			}
+			for _, args := range installer.Spec.SelinuxConfig.SemanageCmds {
+				err := runCommand("semanage", args)
+				if err != nil {
+					return errors.Wrap(err, "failed to run semanage")
+				}
 			}
 		}
 	}
@@ -173,37 +175,39 @@ func processFirewalldConfig(installer *kurlv1beta1.Installer, execCmds bool, gen
 	}
 
 	deleteScript := true
-	if generateScript && installer.Spec.FirewalldConfig.Firewalld != "" {
-		scriptLines := []string{
-			"BYPASS_FIREWALLD_WARNING=1",
+	if installer.Spec.FirewalldConfig != nil {
+		if generateScript && installer.Spec.FirewalldConfig.Firewalld != "" {
+			scriptLines := []string{
+				"BYPASS_FIREWALLD_WARNING=1",
+			}
+
+			switch installer.Spec.FirewalldConfig.Firewalld {
+			case "enabled":
+				scriptLines = append(scriptLines, "systemctl start firewalld")
+				scriptLines = append(scriptLines, "systemctl enable firewalld")
+			case "disabled":
+				scriptLines = append(scriptLines, "if ! systemctl -q is-active firewalld ; then")
+				scriptLines = append(scriptLines, "	return")
+				scriptLines = append(scriptLines, "fi")
+				scriptLines = append(scriptLines, "systemctl stop firewalld")
+				scriptLines = append(scriptLines, "systemctl disable firewalld")
+			default:
+				return errors.Errorf("unknown firewalld option: %s", installer.Spec.FirewalldConfig.Firewalld)
+			}
+
+			script := fmt.Sprintf("configure_firewalld() {\n\t%s\n}", strings.Join(scriptLines, "\n\t"))
+			if err := writeScript(scriptFilename, script); err != nil {
+				return errors.Wrap(err, "faied to save script")
+			}
+			deleteScript = false
 		}
 
-		switch installer.Spec.FirewalldConfig.Firewalld {
-		case "enabled":
-			scriptLines = append(scriptLines, "systemctl start firewalld")
-			scriptLines = append(scriptLines, "systemctl enable firewalld")
-		case "disabled":
-			scriptLines = append(scriptLines, "if ! systemctl -q is-active firewalld ; then")
-			scriptLines = append(scriptLines, "	return")
-			scriptLines = append(scriptLines, "fi")
-			scriptLines = append(scriptLines, "systemctl stop firewalld")
-			scriptLines = append(scriptLines, "systemctl disable firewalld")
-		default:
-			return errors.Errorf("unknown firewalld option: %s", installer.Spec.FirewalldConfig.Firewalld)
-		}
-
-		script := fmt.Sprintf("configure_firewalld() {\n\t%s\n}", strings.Join(scriptLines, "\n\t"))
-		if err := writeScript(scriptFilename, script); err != nil {
-			return errors.Wrap(err, "faied to save script")
-		}
-		deleteScript = false
-	}
-
-	if execCmds {
-		for _, args := range installer.Spec.FirewalldConfig.FirewalldCmds {
-			err := runCommand("firewall-cmd", args)
-			if err != nil {
-				return errors.Wrap(err, "failed to run firewall-cmd")
+		if execCmds {
+			for _, args := range installer.Spec.FirewalldConfig.FirewalldCmds {
+				err := runCommand("firewall-cmd", args)
+				if err != nil {
+					return errors.Wrap(err, "failed to run firewall-cmd")
+				}
 			}
 		}
 	}
@@ -219,11 +223,13 @@ func processFirewalldConfig(installer *kurlv1beta1.Installer, execCmds bool, gen
 }
 
 func processIptablesConfig(installer *kurlv1beta1.Installer, execCmds bool, generateScript bool) error {
-	if execCmds {
-		for _, args := range installer.Spec.IptablesConfig.IptablesCmds {
-			err := runCommand("iptables", args)
-			if err != nil {
-				return errors.Wrap(err, "failed to run iptables")
+	if installer.Spec.IptablesConfig != nil {
+		if execCmds {
+			for _, args := range installer.Spec.IptablesConfig.IptablesCmds {
+				err := runCommand("iptables", args)
+				if err != nil {
+					return errors.Wrap(err, "failed to run iptables")
+				}
 			}
 		}
 	}
